+++ .atomist/editors/AddLocalEditor.ts
@@ -1,3 +1,19 @@
+/*
+ * Copyright © 2017 Atomist, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 import { EditProject } from '@atomist/rug/operations/ProjectEditor'
 import { Project } from '@atomist/rug/model/Project'
 import { Pattern } from '@atomist/rug/operations/RugOperation'
@@ -5,7 +21,7 @@ import { Editor, Parameter, Tags } from '@atomist/rug/operations/Decorators'
 
 import { RugParameters } from './RugParameters'
 
-@Editor("AddLocalEditor", "make an editor for modifying the local project, initiating a rug archive if needed")
+@Editor("AddLocalEditor", "add an editor for modifying the local project, initiating a Rug archive if needed")
 @Tags("rug", "atomist", "typescript")
 export class AddLocalEditor implements EditProject {
 
@@ -22,7 +38,7 @@ export class AddLocalEditor implements EditProject {
         description: "description of the new Rug editor",
         required: false
     })
-    description: string = "An editor for modifying this project";
+    description: string = "an editor for modifying this project";
 
     @Parameter(RugParameters.GroupId)
     groupId: string = "local";